Speedup day12 by a factor of 2 using symmetry
The period of the moons orbits is always even because the moons start out stationary. If you consider the physics going forward an backwards you will find the same points at each timestep until you reach the halfway point which will also be stationary.

To exploit this just check if velocities are 0 and if they are you can multiply the step count and get the total period for an axis.